Checking for likable models is now more robust
`phileo.utils_allowed` handles string arguments, and Model arguments now,
and treats them appropriately. This way, you can pass in either a Model
instance OR a string already in the 'app_name.model' format and both are
handled correctly.

When using the {% likes %} tag, the PHILEO_LIKABLE_MODELS list is
checked before the likes for that model are queried. If the model is not
in PHILEO_LIKABLE_MODELS, that model is skipped.