Add jQuery plugin on page load, not domready
The phileo plugin was added to jQuery in the domready event callback. As
this plugin could be loaded at the bottom of a page - AFTER phileo
widget scripts - it was possible for the phileo plugin to not be
loaded when the widgets were initialized in the domready event. As the
phileo plugin does not need the DOM to be ready to set itself up, the
plugin now loads as the page is loading. This allows phileo widgets to
be inserted before the phileo plugin script and for everything to still
work.